fork a new JVM for REPL/server (Fixes #231)
This allows SBT to persist when the REPL is exited.

You still have to hit enter an extra time for some reason related to
the way the command queue is handed.

Making the server stoppable with a keystroke required a hack to make it
think it’s always running in SBT.