Add scripts to help with git workflow.
Two new scripts, `wip` and `ready`. The first just creates a new wip
branch with the given name and also a remote branch. The second is much
more complicated and handles both going from a wip branch to a pull
request (the PR is only created automatically if `hub` is installed) as
well as updating a revised ready branch.