A few improvements to coverage stuff.
* Add coveralls stats to the README.
* Change `ready` script to run using scoverage.
* Fail the tests if minimum coverage isn’t met

The minimum coverage is currently 57%. Please try to aggressively bump
this up in any PR that you can. Since we currently can’t do coverage on
integration tests, changes to the REPL, evaluator, etc. may cause
coverage to drop, but this will require lowering the minimum % in
build.sbt, so we can keep an eye on when that happens.