be sensitive to the order of fields when checking results
Note: only the order of fields in the top-level object is currently
checked. For example, that means the order of fields inside "value"
is not verified.

Produces a nice message if the fields are correct but out of order.

See #313, which this does not fix (more like makes it actually broken)