Another bit of (optional) `ready` magic.
It can take a while for all the tests and coverage to run during the
ready script. This uses terminal-notifier (`brew install
terminal-notifier`) to pop up a notification when it’s all over so you
don’t have to keep checking back.

Of course, if you don’t have terminal-notifier, it works as before.