respond to PR feedback
Simplified some names and cleaned up much of the Attr-wrangling quite a bit.

Converted two more phases to use the new optimalBoundPhase.

SelectorPhase is still using the old lpBoundPhase because it needs a different
morphism (scanPara2) which the new phase cannot (yet) simulate.

Fix numerous bugs in the new optimalBoundPhase, and changed the type of the
annotation function to allow the monad to be threaded properly. In short,
The phase now maps over the result of each call immediately, so that
evaluation happens only once and then the result is used to annotate the
subexpression. In the previous commit, it was only appearing to work because it
was not properly annotating sub-nodes at all!

A few (surprisingly few) planner tests had their generated names shift a bit as
a result.