simple admin UI
A start, and useful, but not enough to call #479 complete.

Loads and displays the filesystem tree, but does not yet
handle running queries, although the UI is stubbed out.

If there is any configuration problem, it should go straight
to the configuration edit UI and not allow any bad connection
information to be saved.

Connection testing is pretty simple here, and only checks that
an `ls` call returns something.