re-work config UI with more helpful interaction
No more editable table, and no more entering the connection
URI directly. Now the URI is broken out into fields for
editing, with real-time feedback for errors that can be
statically detected. When Save is clicked, a test is
performed, and the action does not continue if any error
occurs.

The SlamData path for a mounting is never shown until there
is more than one mounting. That goes for the overview table,
and also for the edit dialog.

The SlamData port is moved to the bottom of the dialog.

Most of the commonly-used text-editing keystrokes for
Mac OS X are now bound, including cut/copy/paste.

All of the config errors that are easy to provoke are now
detected and a nicer alert message is shown.