use Travis' new Docker platform
sudo is not available, so switched to downloading a mongod
binary and installing it locally.

As of today, Travis' caching actually slows the build down by
a minute or two, compared to downloading the dependencies (and the
mongodb binary) from their usual sources. But it might be worth
trying later on, when all of Travis' users aren't debuggging their
builds.

But the builds are much quicker to start and overall seem to run
about 20-30% faster (8-10 minutes compared to 13-15 minutes previously).