Implement null propagation for JS.
Mongo returns undefined (no field) for `foo.bar.baz` if any
sub-projection is undefined, but any expression other than projection
will result in null (EG, `foo.bar.baz + 4`). We now get those semantics
in JS as well.

JsCore does the right thing automatically, but also created a few helper
functions for JS until everything has been converted to JsCore.

* change $where to take a String, not Code (fixes #528),
* insert a space between unary op and argument (`typeof` is an UnOp),
* tightened up various conditions (`===` instead of `==`)
* add an integration test

The one inconsistency (maybe) is the handling of expressions that are
represented as properties in JS. EG, `length(foo)` would result in null
if it were handled by MongoDB, but because it becomes a
projection (`foo.length`) in JS, it ends up as undefined.