Make all the JsCore operators explicit.
This pushes assignment out of JsCore, and so there is now a safeAssign
function that takes JsCore args but returns Js.Expr. This is about as
much as I could separate this from JsCore, but I think it is actually
the best place for it anyway.

Then there are a bunch of changes just switching to the operator objects
and disambiguating some newly-conflicting names.