report consistent and complete coverage when running all tests
scoverage collects coverage data in the project where the source
resides, regardless of which project’s test is running when it occurs.

That means we were formerly collecting coverage data for a portion of
the tests from the other projects (mainly `it`), non-deterministically.
How much coverage was reported just depended on how many of the `it`
tests happened to have already run at the moment that the `core` tests
finished and the report got generated.

Running `core` last means that all the other projects’ coverage data
has already been collected by the time the report is run, which should
give consistent results, and also ensure that the greatest amount of
coverage is included (win!)

Since we now get coverage for the `it` tests, removed the exclusion for
the MongoDB evaluator.