implement `date` and `time` conversions (Fixes #445)
`Date` and `Time`, as well as `Timestamp` and `Interval`,
all become first-class values in `Data`, so that the front-end
can communicate the semantics unambiguously in the logical plan.

`Date` comparisons get special handling in the planner, since
only timestamp-equivalent values actually exist in MongoDB, but
the assumption is that people are using them to store dates in
some cases, and comparing them involves computing the end-of-day
timestamp.

`Time` values, on the other hand, become simple strings (which
happen to be comparable as is). This meant producing a rather
complicated JS expression, so our `whenDefined` scheme is
significantly beefed up. For example, we now check for the
presence of a method before calling it (with a compound test as in
`(value != null && value.getUTCSeconds != null) ? value.getUTCSeconds() ? undefined`),
but do *not* do that test in cases where it's clear that we
know what we're doing (as in `['CO', 'CA'].indexOf(state)`).

The operators in `MathLib` now accept temporal values where the
result is sensible, which flushed out quite a few bugs and corner
cases in those type checkers. There is probably more to do here,
but the current implementation supports several useful ways of combining
constants and values.

Along the way, fixed some of the worst cases associated with #571, but
there are still likely many combinations of constants and mixed types that
don't work right.