Unify fields that are flattened in both multiple contexts.
Fixes #407.

This replaces the `ShapePreservingBuilder` with a new one that postpones
merging the incoming builders and adds a new `FlatteningBuilder` to
handle both the `$unwind` of `flattenArray` and the `$flatMap` of
`flattenObject` consistently and without throwing away the history of
builders.

The new `ShapePreservingBuilder` allows us to get rid of the complicated
`These` (`\&/`) contents of GroupBuilder, since we no longer have groups
that already have a document and are also having a new field built on
them. This also gets rid of a few bad/impossible states for a
GroupBuilder.

A lot of operations just “pass through” the new
`ShapePreservingBuilder`, being applied to the `src` builder, then
having the SPB wrapped around the result of that.

A few other issues were also fixed here:

* cover more cases in `merge`, `objectConcat`, and `distinctBy` (fixes #384),
* catch a bug where we weren’t rewriting a DocVar (fixes #524),
* implement delegate for `WorkflowBuilder.merge`, and
* statically check if we can `projectField` from a `GroupBuilder`.

One negative side effect is that #503 rears its head a bit more, as we
now manage to build a better `DocBuilder` rather than ending up with a
`$simpleMap` that we later tack a `$project` onto. This affects
`selectArrayElement.test`.