replace `Function.const(...)` and `_ => ...` with `κ(...)
Improves readability slightly. `κ`, because this is the 'K'
of SKI fame, but `K` looks like a type parameter and `k` is just
too common and too indistinct.

'S' and 'I' also implemented, as `σ` and `ɩ`, but they're not
being used anywhere at the moment because, well, S is just plain
not useful, and `ɩ` isn't all that much nicer to read than `identity`.
But it's there, in case there's a name conflict or something.