Disable array_concat.
Fixes #597.

It was breaking queries that threw away the resulting aray, and it
doesn’t work in practice anyway, so it is cauterized to avoid the damage
until we get it working.

Added three integration tests – one that this change fixes, and two more
that were already broken, but at least now are tracked by pending
tests (and #600).