Improve handling of object splicing.
Adds a new SpliceBuilder for objects with partially-unknown structure.

This builds on the fix for #587 – that made those precise queries pass,
but minor changes to them could result in other failures. This gives a
more robust solution.

It also finally eliminates the need for `WorkflowBuilder.merge` to ever
fall back to `Workflow.merge` (although `Workflow.merge` still has a
call site elsewhere, so it’s still not quite dead).