Fix non-top-level flattening.
This applies to all `{*}` as well as `[*]` that end up in a mapReduce.

Previously, we did a shallow copy of the value for each entry in the
flattened object, however if the flattened element was nested (EG,
`foo.bar{*}` or foo[0]{*}`) it would overwrite the original on the first
entry, and then flatten incorrectly for the rest of the object. Now we
do a deep copy, using a function stored in the `mapReduce` `scope`.

Notable changes along the way:

* improve JS block rendering;
* add `Bson.toJs`;
* have `JsExecutor` use `Js` instead of `String`, eliminating fragile
  hand-written JS; and
* add support for `scope` to `mapReduce` operations.