Fix projecting an index on a field in a group.
Fixes #620.

`projectIndex` was not passing operations through GroupBuilder or
ShapePreservingBuilder, creating broken physical plans for the former
and overly-complicated ones for the latter.

Also

* make `expr1` total by allowing `ExprBuilder(ExprBuilder(…))` in the
  case of an expr that has no JS representation on top of a JS expr,
* use `mapExpr` more consistently in the planner, rather than calling
  expr1 directly,
* add a `projectField` case for JS expr that avoids stacked
  ExprBuilders (this showed up after fixing the `projectIndex` case for
  SPBs), and
* tighten up the GroupBuilder pass-though on flattenObject, since
  previously it could ruin plans.

Finally, add an integration test for the GroupBuilder fix, since the
results were previously incorrect, but only a planner test for the
ShapePreservingBuilder fix, since the results were already correct, but
the plan was just too complicated.