handle splicing in objectConcat (fixes #497)
Added a primitive `Splice` expression to JsCore, which encapsulates the
JavaScript machinery as a pure computation. This allows splicing to be
handled by SimpleMap. Also implemented the missing cases in objectConcat
for SpliceBuilders.

That resulted in Simple(Flat)Maps showing up in more places, including after
the unwinds introduced by joins, exposing the fact that successive
SimpleFlatMap ops were not composed in a sensible way. The simplest fix was
extending SimpleFlatMap to handle a list of flattened values. Since
a SimpleFlatMap with an empty list of flattened fields is equivalent to
SimpleMap, just went for it and collapsed the two ops into one, with a
potentially empty list of fields to be flattened.

The new SimpleMap is not as clean and, well, simple, as the old SimpleMap,
but it's more general than the old SimpleFlatMap, so on net it's an improvement.
We could get even more generality by making the "flattening" operation
an arbitrary expression, but it would be tricky to get equally clean, efficient JS.

Slightly improved handling of SimpleMap in the RenderTree instance, making it
more obvious what's what.

Now collapsing ExprBuilders in expr2, so the shape is more
consistent when it arrives in objectConcat. You can still get stacked
ExprBuilders in some cases, however, such as when a JS expression is built on
an ExprOp.