Fix AST rendering (many cases) and parsing (a few) (fixes #430)
Parsing fixes:

1. Accept parens around joins.
2. Allow arbitrary expressions inside `[]` and `{}`.
3. Parse numeric literals with exponents.
4. Rip out half-implemented EXTRACT syntax.

Rendering:

1. Always surround joins with parens.
2. Escape single quotes in string constants.
3. Quote identifiers as needed and escape double quotes in them.

This means `day` and `hour`, etc. are no longer keywords,
at least until we decide to support the awful EXTRACT syntax (see #210).

There were some parsing tests that included EXTRACT, which
I just modified to use `date_part`. As long as we have an
equivalent that works, I didn't see any point in preserving a test
for the syntax we've never supported in the compiler.