array splicing, and `||` and `[]` syntax for arrays and strings (fixes #600)
Add `ArraySpliceBuilder` and cases in `arrayConcat` to handle it. Some
wrangling of types and names in WorkflowBuilder to make this work out.
This fixes a previously pending integration test.

Add `JsCore.SpliceArrays` (and `Splice` becomes `SpliceObjects`).

Add SQL syntax for array constants with `[ ]`, and a `||` operator that
works on both arrays and strings. The operator must be typed to either
Str or an array type during type-checking, so that the compiler can translate
it to either (string) Concat or ArrayConcat. Note: type-checking doesn't
push the inferred type up and down in some cases; see pending test
referring to #637.

Other minor stuff:

Add a (pending) missing test and a fix a broken existing test for math.Add.

Fix the type of empty array constants to be recognizable as an array type
(was previously `Top`, now the slightly overwrought `AnonElem[Bottom]`).
This will do until we have a proper types for arrays and objects.

Fix the label for $simpleMap's in the RenderTree instance.

Increase the coverage minimum to 75%.