eradicate dangling references
There were a handful of situations where values were lost along the way
as queries were planned, generating dangling references in the final plan.
The property-based planner tests now detect this error for all the queries we
generate, and we now generate queries with sorting.

Specifically (and there are new regression tests for each of these):

1. Sorting by a JS expression with filtering present (fixes #651)
2. Implicit grouping, with only reduced projections, with filtering by JS
3. Explicit grouping, with only reduced projections, and filtering present

Along the way, a bunch of things got improved:

Fixed `ExprOp.toJs` to *not* translate `Include`; there's not enough information
in the op, so it needs special handling by the caller.

`optimize` now removes unused fields in `SimpleMap`, when the expression is
a possibly nested `Obj`. Note: it works only if the JsMacro does not inspect
the argument (which is essentially always the case).

Broke out the three parallel cases in `toCollectionBuilder` for
`ShapePreservingBuilder` as functions, making the parallels more visible.
This is where most of the missing rewriting was added. Also added a missing
prefix in the `ExprBuilder`.

`shift` now returns the new base, which is used in one of the cases for SPB.

Fixed the threading of state in `distinctBy` so it uses a consistent set of temp
names, and refactored it into a single `for` comprehension to preserve my sanity.

Deleted the special-case for `ShapePreservingF` in `Workflow.merge` which had the
hack to discard the result and therefore had no valid base on one side. This removes
one dependency on `merge` behavior. This change (along with the above) fixed the
bug, but also resulted in a fair amount of silly nesting (projects where one or more
fields are nested under a tmp, with no other fields at the root). This nesting could
be removed by more intelligent optimization, or maybe by more intelligent combination
of projections.

Note: the other change to some plans is new `$project` ops following `SimpleMap`s to
flatten out the newly-nested shapes. It would be good to inline these in `coalesce`, but that
doesn't work in general because we sometimes need a project in that position to flatten
the implied `value` field, and there's no way to tell if that's the case by looking only
at the ``$project` and its source.

Oh, and I added some files to .gitignore.