switch to http4s (fixes #675)
The services (/metadata, /data, /query, and /compile) are now broken out
as separate HttpService instances, with path prefixes used to map each one
under the right path. This will make it easier to add additional APIs in
the future.

That meant making a couple of very small changes to the tests so that we
hit OPTIONS under a valid path.

One other test change was to make the test for a path with encoded spaces in it
actually check the result.

Other than that, this is just mapping the unfiltered API to http4s's. A couple
of things were missing and required workarounds:

1. `unfiltered.util.Port.any`: finds a "random", unbound port for testing purposes. For now, just using 8888.
2. URL decoding of path segments. Will possibly submit a PR to http4s for this.
3. Handling of paths with trailing slashes. These are swallowed by http4s's Path extractors. For now, just treating paths as directory paths when needed. Submitted a bug to http4s, may need to revisit when it's fixed.
4. Had to re-implement unfiltered's "wait for keystroke" behavior. It's ugly, but seems to work well enough.

http4s pulled in scalaz-stream 0.6a, so I updated our dependency. There was only
one minor code change required to fix a deprecation warning.