respond to PR feedback
Switch to using System.in.available, which happens to signal "closed
stream" the same way it does "no input waiting", which has the
fortuitous side-effect of making the server stay running when you start
it from a script (not a shell). Restore testJar to the way it was
before.

Also properly capture the effectful code there in Task.

Remove a FIXME comment.