Enable warnings for non-Unit-returning statements
Enable wartremover's check for non-Unit-returning
expressions used in statement position. Add a function
`ignore` to make the discarding of values explicit when
we need to do it.

Also enabled scalac's warning for discarded values,
which catches where you produce any value at all but
`Unit` is expected, so scalac says, "great", and just
tosses the value.

Both of those only in the `core` project, and wartremover
only in `compile` (not `test`), because specs2.mutable and
swing both break these rules by their nature.