clean up the RenderedTree type
1. Make `nodeType` the first parameter, and
2. reverse the order of the list, so shared structure can be shared.
3. Make `label` optional, reflecting its use.

This corrects the mistake of putting the nodeType at the end, simply because it was added later, and makes it clearer that the label isn't required. Actually, it would make sense for nodeType to be `NonEmptyList`, but that would be painful, and it's easy enough to provide some default in case it's empty, since it's just being turned into a label string.