Add a static file service.
This serves up URIs under /slamdata relative to the location of the jar. It
also redirects / to /slamdata (assuming no other service matches
the path).

This also adds command-line options to the server (primarily to be able
to choose whether or not starting the server opens a client in the
browser). This meant changing the config file from being the first
argument to being passed with `--config`. I also add a `--port` option
just for convenience.