Allow all `main`s to run without a config file.
This creates an empty config (which means it uses the default
port (20223) and has no mounting).

The admin UI had already allowed this, but we needed to support it for
the Web UI in order to switch the installers. Added it to the REPL for
consistency.

This also gets rid of the case where the Web UI could end up defaulting
to port 8080.