various fixes related to MOVE (fixes #711)
Pass paths through as is, now that http4s preserves trailing slashes.

Treat the destination for MOVE as a directory path if the source is a directory path.

Fix `Path.fileOf`, which was always producing a non-relative path (like `/foo`), which makes no sense. Of course, no one was calling it.

Do something reasonable in GET /metadata with non-dir path (return 200 with an empty object in the body).

Add a test for the error that occurs when you try to move a file on top of an existing file.

Note: it would be great to have unit tests for the API doing the right effects, but that's challenging given that those effects are hidden inside Task[Unit]. I don't want to mock the filesystem statefully, so I'm settling for testing around those effects. Perhaps we need a free algebra for actions on the (generic) filesystem, the same way we need a free algebra for actions on MongoDB specifically.