Produce CSV output with "Accept: text/csv" header (first part of #645)
Uses the first `n` results (currently 1000) to identify columns, so that the entire data set is never loaded into memory at once. Uses the same formatting of values as the REPL and the Admin tool's CSV output.

Because the wrangling of Process was the only tricky bit here, I included a test that effects are properly sequenced (and executed only once!)

Also fixed a couple of bugs:
- the REPL would give an error if every result object was `{}`
- `Data.Set` was not formatted properly in tabular output.