detect CSV format (fixes #645)
Accept CSV-like formats with any of six delimiters, single- or
double-quotes, various line terminators, etc. Attempt to parse a subset
of the input with each format, and choose the format that produces the
most successful parse, with the most likely shape.

Inspired by @jdegoes's `textype` code, but using a different ad hoc
scoring function.

The actual parsing is still done by @tototoshi's parser, but now with a
wrapper to catch exceptions during parsing, mostly when we're
test-parsing with the wrong format.

This should work well for most possible inputs, but may fail in certain
tricky cases, such as when the first 10 records are *very*
unrepresentative of the whole file. Most likely our users will be able
to supply us with examples of tricky cases.