In the API and the MongoDB Filesystem, validate that offset >= 0 and limit > 0
Fixes #751.

Note: offset can be zero (indeed, that's a common offset), but limit cannot (because there's no point in requesting zero results, and because MongoDB wants to interpret that value as "no limit", naturally).

Check in the API so we can show a nice error message, and check in the filesystem because the MongoDB driver doesn't handle this gracefully.

Always use a negative limit value, which gives the driver a hint that we won't request additional results.

Also added an integration test for proper skip/limit behavior of the MongoDB cursor.