Implement partial application and constant folding.
This adds a number of additional “simplifications” to the optimizer. A
few are handled directly, but for Invoke, a new parameter has been added
to Func. It’s done this way because the constant folding should handled
in the Typer function, and it makes sense to keep the other
simplifications near there.

There were a few bugs fixed along the way:

* MongoDbPlanner couldn’t handle constant arrays for sort directions,
* `Take` and `Drop` expected the wrong number of arguments, and
* `Substring` handled indexes inconsistently (see #795).

Many constant folds are currently duplicated in the simplifiers. This is
because I had already implemnted some before I realized the typers
should handle that. I left them in because some of them aren’t yet
reflected in the typers and because we’re not yet applying the folds
from the typers. (Doing that will require adding source locations and
moving the typechecker to the LogicalPlan rather than the SQL AST.)