Handle any error when trying to open the browser.
Failing to open the browser should not cause the server to fail. Instead
we now send a message to stderr acknowledging the error and providing
the URL for the user to use.

This doesn’t fix the underlying issue in #797, but it does at least
allow users to continue working when this fails.