zip up directories for downloading
Fixes #763 (the last checklist item).

Wraps Java's ZipOutputStream so that it can be used to transform the bytes of a scalaz-stream Process.

This is all a bit tricky to test without writing the corresponding wrapping for ZipInputStream, which I did not do. Instead, it tests some small data sets for their exact contents, and tests some large data sets for performance, without trying to do both at once.

Includes a simple test at the API level of the zipped content-type and disposition, which required implementing `ls` properly in the test stub. Also changed the expected behavior of `GET /metadata` with a non-existent directory to match that implementation.