consistently return errors in JSON (fixes #757)
Every error case now has a test that asserts that the response:

1. is in JSON format (with the right content type)
2. contains a field `error`
3. with the expected error message

This uncovered *many* cases where errors were in plain text, or did not have a properly formatted `error` message, or were being trapped in the wrong place, or had an unnecessary stack trace.

Also fixed handling of parse errors in the API and the REPL (a recent regression).

Eliminated curly quotes in responses, because they don't render right in specs2 output.