fix temporary collection leaks
Fix a bug where, whenever a query was run without a specified output path, it would actually be executed twice, and then only the second temporary result would be deleted after the result was returned to the client. This meant that GET queries were taking twice as long as they should, and were leaking one copy of their results.

Fix another bug where a query would fail if the specified ouput collection already existed, and leave behind the temporary result (which fixes #821).

Moved code to recursively list files from the web layer to `Backend.lsAll`.

Related fixes in the REPL:
- implement `env`
- fix parsing and handling of paths for the `[path] := select ...` command