do variable substitution before type analysis
Fixes #769.

This allows variables to contain *any* SQL expression (including some things that don't make sense in strict SQL terms, see below.)

This is simpler than trying to do the substitution after the types are analyzed, so a bunch of code is deleted, along with the tests for the type coercion that's no longer needed.

One feature is lost: you can no longer use an arbitrary string when we can tell from the types that the variable should be a String. This was not actualy all that common, but in any case now you have to explicitly quote strings.

One integration test is added for the obvious use cases of `DATE`, `TIMESTAMP`, and `INTERVAL`.

Note: this permissive approach allows things that SQL does not, such as variables referring to anything other than a constant. For example:
```
💪 $ set foo=pop
💪 $ set bar=pop > 1000
💪 $ select :foo from zips where :bar
```