Improve coverage with law checking.
This required some dependency finagling to get everything working
together. And when I tested the IdHandling ring, it turns out it wasn’t
a ring after all. So I simplified it to a monoid (since `Workflow.merge`
disappeared, we didn’t need the other operation anyway) and that allowed
us to toss Spire for the time being.

I also implemented `Arbitrary[Term[?]]` with `Gen.sized`, allowing
`Arbitrary[Term[LogicalPlan]]` (and other functors) to be simplified to
`Arbitrary[LogicalPlan[?]]`, eliminating some duplication and allowing
for checking of traverse laws.

Also managed to take advantage of shapeless-contrib’s instance
derivation stuff in a few cases.